+++ cypress/integration/pages/app/login.spec.js
@@ -1,20 +1,27 @@
 /// <reference types="cypress" />
-describe('pages/app/login/', () => {
-  it('preencha os campos e vá para a página /app/profile', () => {
-    cy.intercept('https://instalura-api-git-master-omariosouto.vercel.app/api/login')
-      .as('userLogin')
-    cy.visit('/app/login')
-    cy.get('#formCadastro input[name="usuario"]').type('omariosouto')
-    cy.get('#formCadastro input[name="senha"]').type('senhasegura')
-    cy.get('#formCadastro button[type="submit"]').click()
-    cy.url().should('include', 'app/profile')
-    cy.wait('@userLogin')
-      .then((intercept) => {
-        const token = intercept.response.body.data.token
 
-        cy.getCookie('APP_TOKEN')
-          .should('exist')
-          .should('have.property', 'value', token)
-      })
+import LoginScreenPageObject from '../../../../src/components/screens/app/LoginScreen/LoginScreen.pageObject'
+
+describe('/pages/app/login/', () => {
+  describe('when fill and submit a form login request', () => {
+    it('go to the profile page', () => {
+      cy.intercept('https://instalura-api-git-master-omariosouto.vercel.app/api/login')
+        .as('userLogin')
+
+      const loginScreen = new LoginScreenPageObject(cy)
+
+      loginScreen
+        .fillLoginForm({ user: 'omariosouto', password: 'senhasegura' })
+        .submitLoginForm()
+
+      cy.url().should('include', 'app/profile')
+      cy.wait('@userLogin')
+        .then((intercept) => {
+          const token = intercept.response.body.data.token
+          cy.getCookie('APP_TOKEN')
+            .should('exist')
+            .should('have.property', 'value', token)
+        })
+    })
   })
 })
