+++ package.json
@@ -9,7 +9,8 @@
 		"commit": "cz",
 		"lint": "eslint --ignore-path .gitignore .",
 		"custom-script": "node ./script-test.js",
-		"test:integration:open": "cypress open"
+		"test:integration:open": "cypress open",
+		"test:integration": "cypress run --headless"
 	},
 	"dependencies": {
 		"@crello/react-lottie": "^0.0.11",
